+++ Sources/Common/Logging/LogLevel.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public enum LogLevel {
+	case info
+	case warning
+	case error
+}
