+++ Sources/Common/Logging/Loggable.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public protocol Loggable {
+	var message: String { get }
+	var level: LogLevel { get }
+}
