+++ Sources/Common/Logging/Logger.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+// Main interface for the logging. Clients can inject their own implementations
+public protocol Logger {
+	func log(_ loggable: Loggable)
+}
